refactor: 🚚 featuresとentitiesの境目をより明確なものに変更しました
