+++ packages/app/src/entities/board/board.model.ts
