+++ packages/app/src/entities/player/player.model.ts
@@ -1,3 +1,7 @@
 export interface Player {
   readonly name: string;
 }
+
+export interface HumanPlayer extends Player {}
+
+export interface CPUPlayer extends Player {}
